extract BaseMegatronMapper
